+++ core/alarm/src/main/java/com/yapp/alarm/AlarmConstants.kt
@@ -13,4 +13,6 @@ object AlarmConstants {
 
     const val EXTRA_IS_SNOOZED = "com.yapp.orbit.EXTRA_IS_SNOOZED"
     const val EXTRA_IS_DISMISS = "com.yapp.orbit.EXTRA_IS_DISMISS"
+
+    const val WEEK_INTERVAL_MILLIS: Long = 7 * 24 * 60 * 60 * 1000
 }
